Bug 1852533 - Main menu tests TestRail matching
