+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/CollectionTest.kt
@@ -15,6 +15,7 @@ import org.junit.Test
 import org.mozilla.fenix.customannotations.SmokeTest
 import org.mozilla.fenix.helpers.AndroidAssetDispatcher
 import org.mozilla.fenix.helpers.HomeActivityIntentTestRule
+import org.mozilla.fenix.helpers.TestAssetHelper
 import org.mozilla.fenix.helpers.TestAssetHelper.getGenericAsset
 import org.mozilla.fenix.helpers.TestHelper.clickSnackbarButton
 import org.mozilla.fenix.ui.robots.browserScreen
@@ -91,6 +92,19 @@ class CollectionTest {
 
         homeScreen {
             verifyCollectionIsDisplayed(collectionName)
+        }
+    }
+
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2283299
+    @Test
+    fun createFirstCollectionFromMainMenuTest() {
+        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+        }.openThreeDotMenu {
+        }.openSaveToCollection {
+            verifyCollectionNameTextField()
         }
     }
