+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/MainMenuTest.kt
@@ -4,6 +4,7 @@
 
 package org.mozilla.fenix.ui
 
+import androidx.core.net.toUri
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.UiDevice
 import mozilla.components.concept.engine.utils.EngineReleaseChannel
@@ -12,17 +13,20 @@ import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import org.mozilla.fenix.R
 import org.mozilla.fenix.customannotations.SmokeTest
 import org.mozilla.fenix.ext.components
 import org.mozilla.fenix.helpers.AndroidAssetDispatcher
-import org.mozilla.fenix.helpers.HomeActivityTestRule
-import org.mozilla.fenix.helpers.MatcherHelper.itemWithText
-import org.mozilla.fenix.helpers.RecyclerViewIdlingResource
+import org.mozilla.fenix.helpers.HomeActivityIntentTestRule
+import org.mozilla.fenix.helpers.MatcherHelper
 import org.mozilla.fenix.helpers.TestAssetHelper
 import org.mozilla.fenix.helpers.TestHelper
+import org.mozilla.fenix.helpers.TestHelper.assertYoutubeAppOpens
 import org.mozilla.fenix.helpers.TestHelper.runWithCondition
+import org.mozilla.fenix.ui.robots.browserScreen
+import org.mozilla.fenix.ui.robots.clickContextMenuItem
 import org.mozilla.fenix.ui.robots.clickPageObject
+import org.mozilla.fenix.ui.robots.homeScreen
+import org.mozilla.fenix.ui.robots.longClickPageObject
 import org.mozilla.fenix.ui.robots.navigationToolbar
 
 class MainMenuTest {
@@ -30,7 +34,7 @@ class MainMenuTest {
     private lateinit var mockWebServer: MockWebServer
 
     @get:Rule
-    val activityTestRule = HomeActivityTestRule.withDefaultSettingsOverrides()
+    val activityTestRule = HomeActivityIntentTestRule.withDefaultSettingsOverrides()
 
     @Before
     fun setUp() {
@@ -46,9 +50,9 @@ class MainMenuTest {
         mockWebServer.shutdown()
     }
 
-    @SmokeTest
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/233849
     @Test
-    fun verifyPageMainMenuItemsTest() {
+    fun verifyTabMainMenuItemsTest() {
         val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
 
         navigationToolbar {
@@ -59,9 +63,55 @@ class MainMenuTest {
         }
     }
 
+    // Verifies the list of items in the homescreen's 3 dot main menu
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/233848
     @SmokeTest
     @Test
-    fun openMainMenuNewTabItemTest() {
+    fun homeMainMenuItemsTest() {
+        homeScreen {
+        }.openThreeDotMenu {
+            verifyHomeThreeDotMainMenuItems(isRequestDesktopSiteEnabled = false)
+        }.openBookmarks {
+            verifyBookmarksMenuView()
+        }.goBack {
+        }.openThreeDotMenu {
+        }.openHistory {
+            verifyHistoryMenuView()
+        }.goBack {
+        }.openThreeDotMenu {
+        }.openDownloadsManager {
+            verifyEmptyDownloadsList()
+        }.goBack {
+        }.openThreeDotMenu {
+        }.openAddonsManagerMenu {
+            verifyAddonsListIsDisplayed(true)
+        }.goBack {
+        }.openThreeDotMenu {
+        }.openSyncSignIn {
+            verifyTurnOnSyncMenu()
+        }.goBack {}
+        homeScreen {
+        }.openThreeDotMenu {
+        }.openWhatsNew {
+            verifyWhatsNewURL()
+        }.goToHomescreen {
+        }.openThreeDotMenu {
+        }.openHelp {
+            verifyHelpUrl()
+        }.goToHomescreen {
+        }.openThreeDotMenu {
+        }.openCustomizeHome {
+            verifyHomePageView()
+        }.goBackToHomeScreen {
+        }.openThreeDotMenu {
+        }.openSettings {
+            verifySettingsView()
+        }
+    }
+
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2284134
+    @Test
+    fun openNewTabTest() {
         val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
 
         navigationToolbar {
@@ -69,58 +119,31 @@ class MainMenuTest {
         }.openThreeDotMenu {
         }.clickNewTabButton {
             verifySearchView()
+        }.submitQuery("test") {
+            verifyTabCounter("2")
         }
     }
 
+    // Device or AVD requires a Google Services Android OS installation with Play Store installed
+    // Verifies the Open in app button when an app is installed
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/387756
     @SmokeTest
     @Test
-    fun openMainMenuBookmarksItemTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+    fun openInAppFunctionalityTest() {
+        val youtubeURL = "vnd.youtube://".toUri()
 
         navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+        }.enterURLAndEnterToBrowser(youtubeURL) {
+            verifyNotificationDotOnMainMenu()
         }.openThreeDotMenu {
-        }.openBookmarks {
-            verifyBookmarksMenuView()
+        }.clickOpenInApp {
+            assertYoutubeAppOpens()
         }
     }
 
-    @SmokeTest
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2284323
     @Test
-    fun openMainMenuHistoryItemTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.openHistory {
-            verifyHistoryListExists()
-        }
-    }
-
-    @SmokeTest
-    @Test
-    fun openMainMenuAddonsTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.openAddonsManagerMenu {
-            TestHelper.registerAndCleanupIdlingResources(
-                RecyclerViewIdlingResource(
-                    activityTestRule.activity.findViewById(R.id.add_ons_list),
-                    1,
-                ),
-            ) {
-                verifyAddonsItems()
-            }
-        }
-    }
-
-    @SmokeTest
-    @Test
-    fun openMainMenuSyncItemTest() {
+    fun openSyncAndSaveDataTest() {
         val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
 
         navigationToolbar {
@@ -132,167 +155,7 @@ class MainMenuTest {
         }
     }
 
-    @SmokeTest
-    @Test
-    fun openMainMenuFindInPageTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.openFindInPage {
-            verifyFindInPageSearchBarItems()
-        }
-    }
-
-    @SmokeTest
-    @Test
-    fun mainMenuDesktopSiteTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.switchDesktopSiteMode {
-        }.openThreeDotMenu {
-            verifyDesktopSiteModeEnabled(true)
-        }
-    }
-
-    @SmokeTest
-    @Test
-    fun mainMenuReportSiteIssueTest() {
-        runWithCondition(
-            // This test will not run on RC builds because the "Report site issue button" is not available.
-            activityTestRule.activity.components.core.engine.version.releaseChannel !== EngineReleaseChannel.RELEASE,
-        ) {
-            val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-            navigationToolbar {
-            }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-            }.openThreeDotMenu {
-            }.openReportSiteIssue {
-                verifyUrl("webcompat.com/issues/new")
-            }
-        }
-    }
-
-    @SmokeTest
-    @Test
-    fun openMainMenuAddToCollectionTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.openSaveToCollection {
-            verifyCollectionNameTextField()
-        }
-    }
-
-    // Test running on beta/release builds in CI:
-    // caution when making changes to it, so they don't block the builds
-    @SmokeTest
-    @Test
-    fun openMainMenuSettingsItemTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.openSettings {
-            verifySettingsView()
-        }
-    }
-
-    @SmokeTest
-    @Test
-    fun mainMenuShareButtonTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.clickShareButton {
-            verifyShareTabLayout()
-        }
-    }
-
-    @SmokeTest
-    @Test
-    fun mainMenuRefreshButtonTest() {
-        val refreshWebPage = TestAssetHelper.getRefreshAsset(mockWebServer)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(refreshWebPage.url) {
-            mDevice.waitForIdle()
-        }.openThreeDotMenu {
-            verifyThreeDotMenuExists()
-        }.refreshPage {
-            verifyPageContent("REFRESHED")
-        }
-    }
-
-    @SmokeTest
-    @Test
-    fun mainMenuForceRefreshTest() {
-        val refreshWebPage = TestAssetHelper.getRefreshAsset(mockWebServer)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(refreshWebPage.url) {
-            mDevice.waitForIdle()
-        }.openThreeDotMenu {
-            verifyThreeDotMenuExists()
-        }.forceRefreshPage {
-            verifyPageContent("REFRESHED")
-        }
-    }
-
-    @Test
-    fun goBackTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-        val nextWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 2)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-            mDevice.waitForIdle()
-        }.openNavigationToolbar {
-        }.enterURLAndEnterToBrowser(nextWebPage.url) {
-            verifyUrl(nextWebPage.url.toString())
-        }.openThreeDotMenu {
-        }.goToPreviousPage {
-            mDevice.waitForIdle()
-            verifyUrl(defaultWebPage.url.toString())
-        }
-    }
-
-    @Test
-    fun goForwardTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-        val nextWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 2)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-            mDevice.waitForIdle()
-        }.openNavigationToolbar {
-        }.enterURLAndEnterToBrowser(nextWebPage.url) {
-            mDevice.waitForIdle()
-            verifyUrl(nextWebPage.url.toString())
-        }.openThreeDotMenu {
-        }.goToPreviousPage {
-            mDevice.waitForIdle()
-            verifyUrl(defaultWebPage.url.toString())
-        }
-
-        // Re-open the three-dot menu for verification
-        navigationToolbar {
-        }.openThreeDotMenu {
-            verifyThreeDotMenuExists()
-        }.goForward {
-            verifyUrl(nextWebPage.url.toString())
-        }
-    }
-
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/243840
     @Test
     fun findInPageTest() {
         val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 3)
@@ -328,35 +191,164 @@ class MainMenuTest {
         }
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2283303
     @Test
-    fun pdfFindInPageTest() {
-        val genericURL =
-            TestAssetHelper.getGenericAsset(mockWebServer, 3)
+    fun switchDesktopSiteModeOnOffTest() {
+        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
 
         navigationToolbar {
-        }.enterURLAndEnterToBrowser(genericURL.url) {
-            clickPageObject(itemWithText("PDF form file"))
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+        }.openThreeDotMenu {
+        }.switchDesktopSiteMode {
+        }.openThreeDotMenu {
+            verifyDesktopSiteModeEnabled(true)
+        }.switchDesktopSiteMode {
+        }.openThreeDotMenu {
+            verifyDesktopSiteModeEnabled(false)
+        }
+    }
+
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/1314137
+    @Test
+    fun setDesktopSiteBeforePageLoadTest() {
+        val webPage = TestAssetHelper.getGenericAsset(mockWebServer, 4)
+
+        homeScreen {
+        }.openThreeDotMenu {
+            verifyDesktopSiteModeEnabled(false)
+        }.switchDesktopSiteMode {
+        }.openNavigationToolbar {
+        }.enterURLAndEnterToBrowser(webPage.url) {
+        }.openThreeDotMenu {
+            verifyDesktopSiteModeEnabled(true)
+        }.closeBrowserMenuToBrowser {
+            clickPageObject(MatcherHelper.itemContainingText("Link 1"))
+        }.openThreeDotMenu {
+            verifyDesktopSiteModeEnabled(true)
+        }.closeBrowserMenuToBrowser {
+        }.openNavigationToolbar {
+        }.enterURLAndEnterToBrowser(webPage.url) {
+            longClickPageObject(MatcherHelper.itemWithText("Link 2"))
+            clickContextMenuItem("Open link in new tab")
+            TestHelper.clickSnackbarButton("SWITCH")
+        }.openThreeDotMenu {
+            verifyDesktopSiteModeEnabled(false)
+        }
+    }
+
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2283302
+    @Test
+    fun reportSiteIssueTest() {
+        runWithCondition(
+            // This test will not run on RC builds because the "Report site issue button" is not available.
+            activityTestRule.activity.components.core.engine.version.releaseChannel !== EngineReleaseChannel.RELEASE,
+        ) {
+            val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+
+            navigationToolbar {
+            }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+            }.openThreeDotMenu {
+            }.openReportSiteIssue {
+                verifyUrl("webcompat.com/issues/new")
+            }
+        }
+    }
+
+    // Verifies the Add to home screen option in a tab's 3 dot menu
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/410724
+    @SmokeTest
+    @Test
+    fun addPageShortcutToHomeScreenTest() {
+        val website = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+        val shortcutTitle = TestHelper.generateRandomString(5)
+
+        homeScreen {
+        }.openNavigationToolbar {
+        }.enterURLAndEnterToBrowser(website.url) {
+        }.openThreeDotMenu {
+            expandMenu()
+        }.openAddToHomeScreen {
+            clickCancelShortcutButton()
+        }
+
+        browserScreen {
+        }.openThreeDotMenu {
+            expandMenu()
+        }.openAddToHomeScreen {
+            verifyShortcutTextFieldTitle("Test_Page_1")
+            addShortcutName(shortcutTitle)
+            clickAddShortcutButton()
+            clickAddAutomaticallyButton()
+        }.openHomeScreenShortcut(shortcutTitle) {
+            verifyUrl(website.url.toString())
+            verifyTabCounter("1")
+        }
+    }
+
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/329893
+    @SmokeTest
+    @Test
+    fun mainMenuShareButtonTest() {
+        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+        }.openThreeDotMenu {
+        }.clickShareButton {
+            verifyShareTabLayout()
+        }
+    }
+
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/233604
+    @Test
+    fun navigateBackAndForwardTest() {
+        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+        val nextWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 2)
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+            mDevice.waitForIdle()
+        }.openNavigationToolbar {
+        }.enterURLAndEnterToBrowser(nextWebPage.url) {
+            verifyUrl(nextWebPage.url.toString())
+        }.openThreeDotMenu {
+        }.goToPreviousPage {
+            mDevice.waitForIdle()
+            verifyUrl(defaultWebPage.url.toString())
+        }.openThreeDotMenu {
+        }.goForward {
+            verifyUrl(nextWebPage.url.toString())
+        }
+    }
+
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2195819
+    @SmokeTest
+    @Test
+    fun refreshPageButtonTest() {
+        val refreshWebPage = TestAssetHelper.getRefreshAsset(mockWebServer)
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(refreshWebPage.url) {
+            mDevice.waitForIdle()
         }.openThreeDotMenu {
             verifyThreeDotMenuExists()
-            verifyFindInPageButton()
-        }.openFindInPage {
-            verifyFindInPageNextButton()
-            verifyFindInPagePrevButton()
-            verifyFindInPageCloseButton()
-            enterFindInPageQuery("l")
-            verifyFindNextInPageResult("1/2")
-            clickFindInPageNextButton()
-            verifyFindNextInPageResult("2/2")
-            clickFindInPagePrevButton()
-            verifyFindPrevInPageResult("1/2")
-        }.closeFindInPageWithCloseButton {
-            verifyFindInPageBar(false)
+        }.refreshPage {
+            verifyPageContent("REFRESHED")
+        }
+    }
+
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2265657
+    @Test
+    fun forceRefreshPageTest() {
+        val refreshWebPage = TestAssetHelper.getRefreshAsset(mockWebServer)
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(refreshWebPage.url) {
+            mDevice.waitForIdle()
         }.openThreeDotMenu {
-        }.openFindInPage {
-            enterFindInPageQuery("p")
-            verifyFindNextInPageResult("1/1")
-        }.closeFindInPageWithBackButton {
-            verifyFindInPageBar(false)
+            verifyThreeDotMenuExists()
+        }.forceRefreshPage {
+            verifyPageContent("REFRESHED")
         }
     }
 }
