+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/PDFViewerTest.kt
@@ -4,6 +4,7 @@
 
 package org.mozilla.fenix.ui
 
+import androidx.core.net.toUri
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.UiDevice
 import okhttp3.mockwebserver.MockWebServer
@@ -15,10 +16,10 @@ import org.mozilla.fenix.customannotations.SmokeTest
 import org.mozilla.fenix.helpers.AndroidAssetDispatcher
 import org.mozilla.fenix.helpers.Constants.PackageName.GOOGLE_DOCS
 import org.mozilla.fenix.helpers.HomeActivityIntentTestRule
+import org.mozilla.fenix.helpers.MatcherHelper
 import org.mozilla.fenix.helpers.MatcherHelper.itemContainingText
 import org.mozilla.fenix.helpers.MatcherHelper.itemWithResIdAndText
 import org.mozilla.fenix.helpers.MatcherHelper.itemWithText
-import org.mozilla.fenix.helpers.TestAssetHelper
 import org.mozilla.fenix.helpers.TestAssetHelper.getGenericAsset
 import org.mozilla.fenix.helpers.TestHelper.assertExternalAppOpens
 import org.mozilla.fenix.helpers.TestHelper.deleteDownloadedFileOnStorage
@@ -28,6 +29,11 @@ import org.mozilla.fenix.ui.robots.navigationToolbar
 
 class PDFViewerTest {
     private lateinit var mockWebServer: MockWebServer
+    private val downloadTestPage =
+        "https://storage.googleapis.com/mobile_test_assets/test_app/downloads.html"
+    private val pdfFileName = "washington.pdf"
+    private val pdfFileURL = "storage.googleapis.com/mobile_test_assets/public/washington.pdf"
+    private val pdfFileContent = "Washington Crossing the Delaware"
 
     @get:Rule
     val activityTestRule = HomeActivityIntentTestRule.withDefaultSettingsOverrides()
@@ -50,7 +56,7 @@ class PDFViewerTest {
     @Test
     fun openPDFInBrowserTest() {
         val genericURL =
-            TestAssetHelper.getGenericAsset(mockWebServer, 3)
+            getGenericAsset(mockWebServer, 3)
 
         navigationToolbar {
         }.enterURLAndEnterToBrowser(genericURL.url) {
@@ -87,5 +93,56 @@ class PDFViewerTest {
             assertExternalAppOpens(GOOGLE_DOCS)
         }
         deleteDownloadedFileOnStorage(downloadFile)
+    }
+
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2283305
+    @Test
+    fun pdfFindInPageTest() {
+        val genericURL = getGenericAsset(mockWebServer, 3)
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(genericURL.url) {
+            clickPageObject(MatcherHelper.itemWithText("PDF form file"))
+        }.openThreeDotMenu {
+            verifyThreeDotMenuExists()
+            verifyFindInPageButton()
+        }.openFindInPage {
+            verifyFindInPageNextButton()
+            verifyFindInPagePrevButton()
+            verifyFindInPageCloseButton()
+            enterFindInPageQuery("l")
+            verifyFindNextInPageResult("1/2")
+            clickFindInPageNextButton()
+            verifyFindNextInPageResult("2/2")
+            clickFindInPagePrevButton()
+            verifyFindPrevInPageResult("1/2")
+        }.closeFindInPageWithCloseButton {
+            verifyFindInPageBar(false)
+        }.openThreeDotMenu {
+        }.openFindInPage {
+            enterFindInPageQuery("p")
+            verifyFindNextInPageResult("1/1")
+        }.closeFindInPageWithBackButton {
+            verifyFindInPageBar(false)
+        }
+    }
+
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2284297
+    @Test
+    fun addPDFToHomeScreenTest() {
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(downloadTestPage.toUri()) {
+            clickPageObject(MatcherHelper.itemContainingText(pdfFileName))
+            verifyUrl(pdfFileURL)
+            verifyPageContent(pdfFileContent)
+        }.openThreeDotMenu {
+            expandMenu()
+        }.openAddToHomeScreen {
+            verifyShortcutTextFieldTitle(pdfFileName)
+            clickAddShortcutButton()
+            clickAddAutomaticallyButton()
+        }.openHomeScreenShortcut(pdfFileName) {
+            verifyUrl(pdfFileURL)
+        }
     }
 }
