+++ fenix/automation/taskcluster/androidTest/flank-arm-legacy-api-tests.yml
@@ -26,7 +26,6 @@ gcloud:
     - class org.mozilla.fenix.ui.MainMenuTest#goBackTest
     - class org.mozilla.fenix.ui.MainMenuTest#goForwardTest
     - class org.mozilla.fenix.ui.HistoryTest#visitedUrlHistoryTest
-    - class org.mozilla.fenix.ui.MainMenuTest#openMainMenuSettingsItemTest
     - class org.mozilla.fenix.ui.SettingsSearchTest#toggleSearchSuggestionsTest
     - class org.mozilla.fenix.ui.CollectionTest#deleteCollectionTest
     - class org.mozilla.fenix.ui.SmokeTest#noHistoryInPrivateBrowsingTest
