+++ fenix/automation/taskcluster/androidTest/flank-arm-start-test.yml
@@ -21,7 +21,6 @@ gcloud:
 
   test-targets:
    - class org.mozilla.fenix.ui.HistoryTest#visitedUrlHistoryTest
-   - class org.mozilla.fenix.ui.MainMenuTest#openMainMenuSettingsItemTest
    - class org.mozilla.fenix.ui.SettingsSearchTest#toggleSearchSuggestionsTest
    - class org.mozilla.fenix.ui.CollectionTest#deleteCollectionTest
    - class org.mozilla.fenix.ui.SmokeTest#noHistoryInPrivateBrowsingTest
